+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/RunPartitionAliquotDto.java
@@ -4,26 +4,18 @@ public class RunPartitionAliquotDto {
 
   // This ID is only for the sake of DataTables selection and doesn't relate to anything in the DB
   public Long id;
-
   public Long runId;
-
   public String platformType;
-
   public Long containerId;
-
   public String containerIdentificationBarcode;
-
   public Long partitionId;
-
   public Integer partitionNumber;
-
   public Long aliquotId;
-
   public String aliquotName;
-
   public String aliquotAlias;
-
   public Long runPurposeId;
+  public Boolean qcPassed;
+  public String qcNote;
 
   public Long getId() {
     return id;
@@ -111,6 +103,22 @@ public class RunPartitionAliquotDto {
 
   public void setRunPurposeId(Long runPurposeId) {
     this.runPurposeId = runPurposeId;
+  }
+
+  public Boolean getQcPassed() {
+    return qcPassed;
+  }
+
+  public void setQcPassed(Boolean qcPassed) {
+    this.qcPassed = qcPassed;
+  }
+
+  public String getQcNote() {
+    return qcNote;
+  }
+
+  public void setQcNote(String qcNote) {
+    this.qcNote = qcNote;
   }
 
 }