+++ sqlstore/src/main/resources/db/migration/V8501__run_library_qc.sql
@@ -1,0 +1,2 @@
+ALTER TABLE Run_Partition_LibraryAliquot ADD COLUMN qcPassed BOOLEAN;
+ALTER TABLE Run_Partition_LibraryAliquot ADD COLUMN qcNote varchar(255);